+++ example/lib/src/home.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:universe/universe.dart';
-import 'package:universe_example/src/shared.dart';
 
 class HomePage extends StatefulWidget {
   @override
@@ -11,8 +10,7 @@ class HomePage extends StatefulWidget {
 class _HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
-    return U.MapBox(
-      accessToken: yourAccessTokenHere,
+    return U.GoogleMap(
       center: [-6.175329, 106.827253],
       zoom: 15,
       onReady: () => print('Map ready!'),